requirements.txt
# Reddit
praw
# Python environment
python-dotenv